PathMatcher : Declare hash function overload for PathMatcher
