+++ src/IECore/PathMatcher.cpp
@@ -39,6 +39,8 @@
 
 #include "IECore/StringAlgo.h"
 
+#include <stack>
+
 using namespace std;
 using namespace IECore;
 
@@ -239,11 +241,11 @@ unsigned PathMatcher::match( const std::string &path ) const
 
 unsigned PathMatcher::match( const std::vector<IECore::InternedString> &path ) const
 {
-   Node *node = m_root.get();
-   if( !node )
-   {
-       return NoMatch;
-   }
+	Node *node = m_root.get();
+	if( !node )
+	{
+		return NoMatch;
+	}
 
 	unsigned result = NoMatch;
 	matchWalk( node, path.begin(), path.end(), result );
@@ -750,4 +752,84 @@ PathMatcher::NodePtr PathMatcher::removePathsWalk( Node *node, const Node *srcNo
 	}
 
 	return result;
+}
+
+namespace
+{
+
+//////////////////////////////////////////////////////////////////////////
+// Support code for murmurHashAppend for PathMatcher
+//////////////////////////////////////////////////////////////////////////
+
+struct HashNode
+{
+
+	HashNode( const char *name, unsigned char exactMatch )
+		:   name( name ), exactMatch( exactMatch )
+	{
+	}
+
+	bool operator < ( const HashNode &rhs ) const
+	{
+		return strcmp( name, rhs.name ) < 0;
+	}
+
+	const char *name;
+	unsigned char exactMatch;
+
+};
+
+typedef std::vector<HashNode> HashNodes;
+typedef std::stack<HashNodes> HashStack;
+
+void popHashNodes( HashStack &stack, size_t size, IECore::MurmurHash &h )
+{
+	while( stack.size() > size )
+	{
+		h.append( (uint64_t)stack.top().size() );
+		std::sort( stack.top().begin(), stack.top().end() );
+		for( HashNodes::const_iterator nIt = stack.top().begin(), nEIt = stack.top().end(); nIt != nEIt; ++nIt )
+		{
+			h.append( nIt->name );
+			h.append( nIt->exactMatch );
+		}
+		stack.pop();
+	}
+}
+
+}
+
+// Our hash is complicated by the fact that PathMatcher::Iterator doesn't
+// guarantee the order of visiting child nodes in its tree (because it
+// sorts using InternedString addresses for the fastest possible match()
+// implementation). We therefore have to use a stack to keep track of
+// our traversal through the tree, and output all the children at each
+// level only after sorting them alphabetically.
+void IECore::murmurHashAppend( IECore::MurmurHash &h, const IECore::PathMatcher &data )
+{
+	HashStack stack;
+	for( PathMatcher::RawIterator it = data.begin(), eIt = data.end(); it != eIt; ++it )
+	{
+		// The iterator is recursive, so we use a stack to keep
+		// track of where we are. Resize the stack to match our
+		// current depth. The required size has the +1 because
+		// we need a stack entry for the root item.
+		size_t requiredStackSize = it->size() + 1;
+		if( requiredStackSize > stack.size() )
+		{
+			// Going a level deeper.
+			stack.push( HashNodes() );
+			assert( stack.size() == requiredStackSize );
+		}
+		else if( requiredStackSize < stack.size() )
+		{
+			// Returning from recursion to the child nodes.
+			// Output the hashes for the children we visited
+			// and stored on the stack previously.
+			popHashNodes( stack, requiredStackSize, h );
+		}
+
+		stack.top().push_back( HashNode( it->size() ? it->back().c_str() : "", it.exactMatch() ) );
+	}
+	popHashNodes( stack, 0, h );
 }
